contractcourt: move responsibility for closing force closes to chanArb
This commit moves the responsibility for closing local and remote force
closes in the database from the chain watcher to the channel arbitrator.

We do this because we previously would close the channel in the
database, before sending the event to the channel arbitrator. This could
lead to a situation where the channel was marked closed, but the channel
arbitrator didn't receive the event before shutdown. As we don't listen
for chain events for channels that are closed, those channels would be
stuck in the pending close state forever, as the channel arbitrator
state machine wouldn't progress.

We introduce a new state StateCommitmentConfirmed that is used
exclusively for marking the channel closed in the database. Before
transitioning to this new state, we make sure the ChannelCloseSummary is
written to the log in the stateCallback. And before transitioning away
from this state (to StateContractClosed) we make sure the channel is
marked closed in the database. This ensures that if any of these steps
fail, we have the information necessary to recover.